rework, but now somehow my complexity in the 4wide puzzle is borked
